chore: addressing feedback comments
- Protobuf Row to Result is now private method instead of class.
- removed misleading comment on timestamp dedup.
